+++ pytorch_lightning/accelerators/legacy/cpu_accelerator.py
@@ -15,43 +15,38 @@ from typing import Any, Callable, Optional, Union
 
 import torch
 
-from pytorch_lightning.accelerators.accelerator import Accelerator, ReduceOp
+from pytorch_lightning.accelerators.legacy.accelerator import Accelerator, ReduceOp
 from pytorch_lightning.cluster_environments import ClusterEnvironment
-from pytorch_lightning.distributed.dist import LightningDistributed
 from pytorch_lightning.utilities import AMPType
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 
-class GPUAccelerator(Accelerator):
-    amp_backend: AMPType
+class CPUAccelerator(Accelerator):
 
     def __init__(self, trainer, cluster_environment: Optional[ClusterEnvironment] = None):
         """
-        Runs training using a single GPU
+        Runs training on CPU
 
         Example::
 
             # default
-            trainer = Trainer(accelerator=GPUAccelerator())
+            trainer = Trainer(accelerator=CPUAccelerator())
 
         """
         super().__init__(trainer, cluster_environment)
-        self.dist = LightningDistributed()
         self.nickname = None
 
     def setup(self, model):
+        # run through amp wrapper
+        if self.trainer.amp_backend:
+            raise MisconfigurationException('amp + cpu is not supported.  Please use a GPU option')
 
-        # call setup
+        # call setup after the ddp process has connected
         self.trainer.call_setup_hook(model)
-
-        torch.cuda.set_device(self.trainer.root_gpu)
-        model.cuda(self.trainer.root_gpu)
 
         # CHOOSE OPTIMIZER
         # allow for lr schedulers as well
         self.setup_optimizers(model)
-
-        # 16-bit
-        model = self.trainer.precision_connector.connect(model)
 
         self.trainer.model = model
 
@@ -66,14 +61,11 @@ class GPUAccelerator(Accelerator):
         return results
 
     def _step(self, model_step: Callable, args):
-        args[0] = self.to_device(args[0])
-
         if self.trainer.amp_backend == AMPType.NATIVE:
             with torch.cuda.amp.autocast():
                 output = model_step(*args)
         else:
             output = model_step(*args)
-
         return output
 
     def training_step(self, args):
@@ -84,16 +76,6 @@ class GPUAccelerator(Accelerator):
 
     def test_step(self, args):
         return self._step(self.trainer.model.test_step, args)
-
-    def to_device(self, batch):
-        gpu_id = 0
-        if isinstance(self.trainer.data_parallel_device_ids, list):
-            gpu_id = self.trainer.data_parallel_device_ids[0]
-
-        # Don't copy the batch since there is a single gpu that the batch could
-        # be referenced from and if there are multiple optimizers the batch will
-        # wind up copying it to the same device repeatedly.
-        return self.batch_to_device(batch, gpu_id)
 
     def sync_tensor(self,
                     tensor: Union[torch.Tensor],
